move create appointment type to type folder
